add bulk delete all and copy all for filtered photos

+++ service/deleter.go
@@ -14,11 +14,15 @@ func NewDeleter() *Deleter {
 }
 
 func (d *Deleter) Delete(photo model.Photo) error {
+	return d.DeleteWithOption(photo, true)
+}
+
+func (d *Deleter) DeleteWithOption(photo model.Photo, includeRAW bool) error {
 	if err := os.Remove(photo.ImagePath); err != nil && !os.IsNotExist(err) {
 		return fmt.Errorf("deleting image %s: %w", photo.ImagePath, err)
 	}
 
-	if photo.HasRAW() {
+	if includeRAW && photo.HasRAW() {
 		if err := os.Remove(photo.RAWPath); err != nil && !os.IsNotExist(err) {
 			return fmt.Errorf("deleting RAW %s: %w", photo.RAWPath, err)
 		}
